Some initial grouping of the refactored MathTrig functions into small group classes, and renaming of methods
