--- src/PhpSpreadsheet/Calculation/MathTrig/FloorPrecise.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace PhpOffice\PhpSpreadsheet\Calculation\MathTrig;
-
-use Exception;
-use PhpOffice\PhpSpreadsheet\Calculation\Functions;
-
-class FloorPrecise
-{
-    /**
-     * FLOOR.PRECISE.
-     *
-     * Rounds number down, toward zero, to the nearest multiple of significance.
-     *
-     * Excel Function:
-     *        FLOOR.PRECISE(number[,significance])
-     *
-     * @param float $number Number to round
-     * @param float $significance Significance
-     *
-     * @return float|string Rounded Number, or a string containing an error
-     */
-    public static function evaluate($number, $significance = 1)
-    {
-        try {
-            $number = Helpers::validateNumericNullBool($number);
-            $significance = Helpers::validateNumericNullSubstitution($significance, null);
-        } catch (Exception $e) {
-            return $e->getMessage();
-        }
-
-        return self::argumentsOk((float) $number, (float) $significance);
-    }
-
-    /**
-     * Avoid Scrutinizer problems concerning complexity.
-     *
-     * @return float|string
-     */
-    private static function argumentsOk(float $number, float $significance)
-    {
-        if ($significance == 0.0) {
-            return Functions::DIV0();
-        }
-        if ($number == 0.0) {
-            return 0.0;
-        }
-
-        return floor($number / abs($significance)) * abs($significance);
-    }
-}
